# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api de Atendimento",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/"
        }
    ],
    "paths": {
...
        "/servicos/find/{servico}": {
            "get": {
                "tags": [
                    "/servicos/find/:servico"
                ],
                "summary": "Pega serviço por Id",
                "description": "Retorna serviço por Id",
                "operationId": "App\\Http\\Controllers\\ServicoController::show",
                "parameters": [
                    {
                        "name": "servico",
                        "in": "path",
                        "description": "servico id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso!"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/servicos/create": {
            "post": {
                "tags": [
                    "/servicos/create"
                ],
                "summary": "Cria um novo Serviço",
                "description": "Retorna um novo Serviço",
                "operationId": "App\\Http\\Controllers\\ServicoController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "qtd_min": {
                                        "type": "integer"
                                    },
                                    "valor": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucesso!"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/servicos/update/{servico}": {
            "put": {
                "tags": [
                    "/servicos/update/:servico"
                ],
                "summary": "Atualiza serviço existente",
                "description": "Retorna serviço atualizado por id",
                "operationId": "App\\Http\\Controllers\\ServicoController::update",
                "parameters": [
                    {
                        "name": "servico",
                        "in": "path",
                        "description": "servico id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string"
                                    },
                                    "qtd_min": {
                                        "type": "integer"
                                    },
                                    "valor": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Sucesso"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/servicos/{servico}": {
            "delete": {
                "tags": [
                    "/servicos/:servico"
                ],
                "summary": "Deleta um Serviço existente",
                "description": "Deleta o Serviço por Id",
                "operationId": "App\\Http\\Controllers\\ServicoController::delete",
                "parameters": [
                    {
                        "name": "servico",
                        "in": "path",
                        "description": "servico id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/servicos/addFuncionario/{servico}": {
            "post": {
                "tags": [
                    "/servicos/addFuncionario/:servico"
                ],
                "summary": "Adiciona funcionario ao servico",
                "description": "Adiciona funcionario ao servico",
                "operationId": "App\\Http\\Controllers\\ServicoController::storeFuncionario",
                "parameters": [
                    {
                        "name": "servico",
                        "in": "path",
                        "description": "servico id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "funcionario_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucesso!"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/upload": {
            "post": {
                "tags": [
                    "/upload"
                ],
                "summary": "Faz Upload de Imagem",
                "description": "Faz Upload de Imagem",
                "operationId": "App\\Http\\Controllers\\UploadController::uploadImage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "type": "file",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucesso!"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}